moved Aggregator class to correct package
