+++ health/src/main/java/com/health/operations/Aggregator.java
@@ -1,4 +1,4 @@
-package com.health.input;
+package com.health.operations;
 
 import com.health.AggregateFunctions;
